+++ packages/x-ray/__test__/index.test.ts
@@ -11,7 +11,6 @@ describe('bearerOverride', () => {
     bearerOverride()
 
     expect(require('http')._bearerLoading).toBeDefined()
-    expect(require('https')._bearerLoading).toBeDefined()
     expect(require('console')._bearerLoading).toBeDefined()
   })
 })