+++ packages/x-ray/src/index.ts
@@ -6,22 +6,20 @@ let check: string
 
 export const bearerOverride = () => {
   const httpModule = require('http')
-  const httpsModule = require('https')
+  // const httpsModule = require('https')
   const consoleModule = require('console')
 
-  if (httpModule._bearerLoading === check && httpsModule._bearerLoading && consoleModule._bearerLoading && check) {
+  if (httpModule._bearerLoading === check && consoleModule._bearerLoading && check) {
     logger('%j', { message: 'modules have already been loaded', application: 'x-ray' })
     return
   }
 
-  httpsModule._bearerLoading = httpModule._bearerLoading = consoleModule._bearerLoading = check = [
-    Math.random(),
-    Date.now()
-  ].join('/')
+  httpModule._bearerLoading = consoleModule._bearerLoading = check = [Math.random(), Date.now()].join('/')
   logger('%j', { message: 'Overriding request and get methods', application: 'x-ray' })
 
   overrideRequestMethod(httpModule)
-  overrideRequestMethod(httpsModule)
+  // AWS does the stuff for us
+  // overrideRequestMethod(httpsModule)
   overrideConsole(consoleModule)
 }