+++ packages/cli/CHANGELOG.md
@@ -3,6 +3,33 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.6.0](https://github.com/Bearer/bearer/compare/v1.5.0...v1.6.0) (2019-07-02)
+
+**Note:** Version bump only for package @bearer/cli
+
+
+
+
+
+# [1.5.0](https://github.com/Bearer/bearer/compare/v1.4.1...v1.5.0) (2019-07-02)
+
+
+### Bug Fixes
+
+* add more fields to logs ([50b6c96](https://github.com/Bearer/bearer/commit/50b6c96))
+* improve integration not found error message ([06207cb](https://github.com/Bearer/bearer/commit/06207cb))
+* revert cli update ([368a7c5](https://github.com/Bearer/bearer/commit/368a7c5))
+* **cli:** use refresh token for every dev portal request ([#695](https://github.com/Bearer/bearer/issues/695)) ([c913308](https://github.com/Bearer/bearer/commit/c913308))
+
+
+### Features
+
+* **x-ray:** drop more information into logs ([90c35e8](https://github.com/Bearer/bearer/commit/90c35e8))
+
+
+
+
+
 ## [1.4.1](https://github.com/Bearer/bearer/compare/v1.4.0...v1.4.1) (2019-06-21)