+++ packages/node/CHANGELOG.md
@@ -3,6 +3,33 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.6.0](https://github.com/Bearer/bearer/compare/v1.5.0...v1.6.0) (2019-07-02)
+
+
+### Features
+
+* **node:** multi-method client ([#700](https://github.com/Bearer/bearer/issues/700)) ([d9ef122](https://github.com/Bearer/bearer/commit/d9ef122))
+
+
+
+
+
+# [1.5.0](https://github.com/Bearer/bearer/compare/v1.4.1...v1.5.0) (2019-07-02)
+
+
+### Bug Fixes
+
+* **node:** direct link to developer keys ([d7e1bd0](https://github.com/Bearer/bearer/commit/d7e1bd0))
+
+
+### Features
+
+* **node:** raise error if apiKey is not provided ([6fcc6cd](https://github.com/Bearer/bearer/commit/6fcc6cd))
+
+
+
+
+
 ## [0.115.1](https://github.com/Bearer/bearer/compare/v0.115.0...v0.115.1) (2019-04-16)
 
 **Note:** Version bump only for package @bearer/node