+++ packages/openapi-generator/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.6.0](https://github.com/Bearer/bearer/compare/v1.5.0...v1.6.0) (2019-07-02)
+
+**Note:** Version bump only for package @bearer/openapi-generator
+
+
+
+
+
+# [1.5.0](https://github.com/Bearer/bearer/compare/v1.4.1...v1.5.0) (2019-07-02)
+
+**Note:** Version bump only for package @bearer/openapi-generator
+
+
+
+
+
 # [1.4.0](https://github.com/Bearer/bearer/compare/v1.3.1...v1.4.0) (2019-06-20)
 
 **Note:** Version bump only for package @bearer/openapi-generator